Rewrote and fixed stocks per #193. Read on.
Stocks has been rewritten to rely exclusively on two tools: jq and
curl. Other request tools have been removed. This simplifies the
code and makes it easier to support. The dependency on Python 2
was also removed from Stocks. Python 2 support is set to expire
soon, and using jq in it's place cleans up the code mightily.

Update and argument flag support has been removed. the Update
command is really outside the scope of stocks and should be
its own command or sourced as a support script. The argument
flags were just extra complication. There is now one, and only
one way to invoke a command.

The old IEX API was deprecated. This now points to the new clouds
API. Unfortunately, that API requires a token. The script now
looks for STOCKSIEXTOKEN. I don't know if checking the token
in the environment is the right call, but it solves the problem
for me. Users can make their own decisions.

The script has been hardened against some word splitting and
some documentation cleaned up. The tests were also changed
to reflect the changes.